requirements-dev.txt
# GLOBAL REQUIREMENTS.
-r requirements.txt

# SOFT DEPENDENCIES
tqdm
h5py
scooby
xarray
empymod
discretize
matplotlib

# SETUP RELATED
setuptools_scm
mkl-service

# FOR DOCUMENTATION
pooch
sphinx > 3
numpydoc >= 0.9
sphinx_gallery
sphinx_panels
sphinx_numfig
pydata_sphinx_theme
sphinx_automodapi
ipykernel

# FOR TESTING
asv
pytest
coveralls
pytest-cov
pytest-flake8
pytest-console-scripts
